api/cs415/cs415/requirements.txt
asgiref==3.7.2
Django==4.2.9
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
mysqlclient==2.1.1
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
